src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
beautifulsoup4==4.12.3
    # via newswatch (pyproject.toml)
boto3==1.34.162
    # via newswatch (pyproject.toml)
botocore==1.34.162
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via nltk
google-api-core==2.24.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.25.0
    # via newswatch (pyproject.toml)
google-cloud-core==2.4.2
    # via google-cloud-bigquery
google-crc32c==1.6.0
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.68.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via nltk
nltk==3.9.1
    # via
    #   newswatch (pyproject.toml)
    #   textblob
packaging==24.2
    # via google-cloud-bigquery
proto-plus==1.26.0
    # via google-api-core
protobuf==5.29.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==19.0.1
    # via newswatch (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==1.10.21
    # via newswatch (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   google-cloud-bigquery
pyyaml==6.0.2
    # via newswatch (pyproject.toml)
regex==2024.11.6
    # via nltk
requests==2.32.3
    # via
    #   newswatch (pyproject.toml)
    #   google-api-core
    #   google-cloud-bigquery
rsa==4.9
    # via google-auth
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
textblob==0.17.1
    # via newswatch (pyproject.toml)
tqdm==4.67.1
    # via nltk
typing-extensions==4.12.2
    # via pydantic
urllib3==2.3.0
    # via
    #   botocore
    #   requests
